# documentacion/DOCUMENTACION_PAWMILY/03_Manuales_Tecnicos_y_DevOps/Especificacion_APIs_OpenAPI.yaml
openapi: 3.0.3
info:
  title: PawMily API
  version: "1.0.0"
  description: |
    API REST de PawMily (Express Clean Architecture + DDD).
    Base producción: https://api-production-66b1.up.railway.app/api
  contact:
    name: Equipo PawMily
servers:
  - url: https://api-production-66b1.up.railway.app/api
    description: Producción Railway
  - url: http://127.0.0.1:3000/api
    description: Desarrollo local
...
paths:
...
  /patients:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Listar pacientes (vet)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200": { description: data + meta }
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Crear paciente (vet)
      responses:
        "201": { description: Patient }
  /patients/mine:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Mascotas del owner
      responses:
        "200": { description: Lista lean }
  /patients/code/{code}:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Obtener por código PAW
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        "200": { description: Patient }
  /patients/link:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Vincular (legacy/helper)
      responses:
        "200": { description: Patient o status }
  /patients/link-requests:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Crear solicitud de vínculo por código
      responses:
        "201": { description: LinkRequest }
  /patients/link-requests/pending:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Solicitudes pendientes
      responses:
        "200": { description: Lista }
  /patients/link-requests/{id}/approve:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Aprobar solicitud
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: LinkRequest }
  /patients/link-requests/{id}/reject:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Rechazar solicitud
      responses:
        "200": { description: LinkRequest }
  /patients/{id}:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Detalle paciente
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Patient }
    put:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Actualizar paciente
      responses:
        "200": { description: Patient }
    delete:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Eliminar paciente
      responses:
        "200": { description: OK }
  /patients/{id}/barcode:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Datos de código de barras
      responses:
        "200": { description: code/format/imageUrl }
  /patients/{id}/medical-records:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Historial médico
      responses:
        "200": { description: Lista }
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Crear registro médico (vet)
      responses:
        "201": { description: MedicalRecord }
  /patients/{id}/feeding:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Obtener alimentación
      responses:
        "200": { description: Feeding }
    put:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Actualizar feeding
      responses:
        "200": { description: Feeding }
  /patients/{id}/diet:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Definir dieta (vet)
      responses:
        "200": { description: Feeding }
  /patients/{id}/reminders:
    get:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Listar recordatorios
      responses:
        "200": { description: Lista }
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Crear recordatorio
      responses:
        "201": { description: Reminder }
  /patients/reminders/{id}:
    put:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Actualizar recordatorio
      responses:
        "200": { description: Reminder }
  /patients/reminders/{id}/complete:
    post:
      tags: [Patients]
      security: [{ bearerAuth: [] }]
      summary: Completar recordatorio
      responses:
        "200": { description: Reminder }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...